+++ src/lib/opengpex/plugins/base/overlays/FileLoader/components.tsx
@@ -94,7 +94,7 @@ export function FileLoaderComponent() {
           }
           e.target.value = "";
         }}
-        accept="image/*"
+        accept="image/*,.eps,.epsf,application/postscript"
         multiple
         className="hidden"
       />